fix: centralize gateway token counting
