+++ internal/tokencounting/client_test.go
@@ -18,47 +18,6 @@ import (
 
 const bufSize = 1024 * 1024
 
-const tokenCountingGatewayServiceName = "agynio.api.gateway.v1.TokenCountingGateway"
-
-type tokenCountingGatewayServer interface {
-	CountTokens(context.Context, *tokencountingv1.CountTokensRequest) (*tokencountingv1.CountTokensResponse, error)
-}
-
-var tokenCountingGatewayServiceDesc = grpc.ServiceDesc{
-	ServiceName: tokenCountingGatewayServiceName,
-	HandlerType: (*tokenCountingGatewayServer)(nil),
-	Methods: []grpc.MethodDesc{
-		{
-			MethodName: "CountTokens",
-			Handler:    tokenCountingGatewayCountTokensHandler,
-		},
-	},
-	Streams:  []grpc.StreamDesc{},
-	Metadata: "agynio/api/gateway/v1/token_counting.proto",
-}
-
-func registerTokenCountingGatewayServer(s grpc.ServiceRegistrar, srv tokenCountingGatewayServer) {
-	s.RegisterService(&tokenCountingGatewayServiceDesc, srv)
-}
-
-func tokenCountingGatewayCountTokensHandler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (any, error) {
-	in := new(tokencountingv1.CountTokensRequest)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(tokenCountingGatewayServer).CountTokens(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: tokenCountingGatewayCountTokensMethod,
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(tokenCountingGatewayServer).CountTokens(ctx, req.(*tokencountingv1.CountTokensRequest))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
 type captureServer struct {
 	mu   sync.Mutex
 	req  *tokencountingv1.CountTokensRequest
@@ -88,7 +47,7 @@ func newBufconnClient(t *testing.T, server *captureServer) (*Client, func()) {
 	t.Helper()
 	listener := bufconn.Listen(bufSize)
 	grpcServer := grpc.NewServer()
-	registerTokenCountingGatewayServer(grpcServer, server)
+	RegisterTokenCountingGatewayServer(grpcServer, server)
 	go func() {
 		_ = grpcServer.Serve(listener)
 	}()
