+++ test/e2e/token_counting_server_test.go
@@ -13,6 +13,7 @@ import (
 	"sync"
 	"testing"
 
+	"github.com/agynio/agn-cli/internal/tokencounting"
 	tokencountingv1 "github.com/agynio/agn-cli/internal/tokencounting/token_countingv1"
 	"google.golang.org/grpc"
 )
@@ -22,50 +23,6 @@ var (
 	tokenCountingAddr string
 	tokenCountingErr  error
 )
-
-const (
-	tokenCountingGatewayServiceName       = "agynio.api.gateway.v1.TokenCountingGateway"
-	tokenCountingGatewayCountTokensMethod = "/agynio.api.gateway.v1.TokenCountingGateway/CountTokens"
-)
-
-type tokenCountingGatewayServer interface {
-	CountTokens(context.Context, *tokencountingv1.CountTokensRequest) (*tokencountingv1.CountTokensResponse, error)
-}
-
-var tokenCountingGatewayServiceDesc = grpc.ServiceDesc{
-	ServiceName: tokenCountingGatewayServiceName,
-	HandlerType: (*tokenCountingGatewayServer)(nil),
-	Methods: []grpc.MethodDesc{
-		{
-			MethodName: "CountTokens",
-			Handler:    tokenCountingGatewayCountTokensHandler,
-		},
-	},
-	Streams:  []grpc.StreamDesc{},
-	Metadata: "agynio/api/gateway/v1/token_counting.proto",
-}
-
-func registerTokenCountingGatewayServer(s grpc.ServiceRegistrar, srv tokenCountingGatewayServer) {
-	s.RegisterService(&tokenCountingGatewayServiceDesc, srv)
-}
-
-func tokenCountingGatewayCountTokensHandler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (any, error) {
-	in := new(tokencountingv1.CountTokensRequest)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(tokenCountingGatewayServer).CountTokens(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: tokenCountingGatewayCountTokensMethod,
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(tokenCountingGatewayServer).CountTokens(ctx, req.(*tokencountingv1.CountTokensRequest))
-	}
-	return interceptor(ctx, in, info, handler)
-}
 
 func tokenCountingAddress(t *testing.T) string {
 	t.Helper()
@@ -84,7 +41,7 @@ func startTokenCountingServer() (string, error) {
 		return "", err
 	}
 	server := grpc.NewServer()
-	registerTokenCountingGatewayServer(server, tokenCountingServer{})
+	tokencounting.RegisterTokenCountingGatewayServer(server, tokenCountingServer{})
 	go func() {
 		_ = server.Serve(listener)
 	}()
